Rename freepybox to aiofreepybox
