+++ aiofreepybox/exceptions.py
